Update post router to respond with user profile
- separate the function of getting url after uploading image
- update sequelize to include user profile in response
